sandbox/river-train/requirements.txt
scikit-learn==0.23.2
progressbar2==3.53.1